move reducers to settings section
